requirements.txt
selenium
nose
nose-testconfig
